feat: property api index (#1)
* feat: property api index

* feat: show property api

* Update app/models/unit_repository.rb


* Update app/controllers/api/v1/properties_controller.rb


* Update app/models/unit_repository.rb


* Update config/routes.rb


* Update app/models/unit.rb


* fix: changes PR

* fix: more changes PR
