--- app/controllers/application_controller.rb
@@ -1,23 +1,0 @@
-class ApplicationController < ActionController::Base
-  protect_from_forgery with: :exception
-
-  helper :all
-  helper_method :home?
-  helper_method :property_detail?
-
-  before_action :weather
-
-  private
-
-  def weather
-    @weather = Weather.get
-  end
-
-  def home?
-    controller_name == 'home' && action_name == 'index'
-  end
-
-  def property_detail?
-    controller_name == 'property' && action_name == 'view'
-  end
-end
