--- app/views/api/shared/_unit.json.jbuilder
@@ -1,10 +1,0 @@
-json.code unit.code
-json.name unit.name
-json.type unit.type
-json.address unit.address
-json.occupancy unit.occupancy
-json.amenities unit.amenities
-json.num_floors unit.num_floors
-json.bedrooms unit.bedrooms
-json.bathrooms unit.bathrooms
-json.descriptions unit.descriptions
