--- app/views/api/units/index.json.jbuilder
@@ -1,5 +1,0 @@
-json.units do
-  json.array! @units do |unit|
-    json.partial! 'api/shared/unit', unit: unit
-  end
-end
