--- app/views/api/units/show.json.jbuilder
@@ -1,3 +1,0 @@
-json.unit do
-  json.partial! 'api/shared/unit', unit: @unit
-end
