+++ app/views/api/v1/home/index.json.jbuilder
@@ -1,0 +1,5 @@
+json.units do
+  json.array! @random_units do |unit|
+    json.partial! 'api/v1/shared/unit', unit: unit
+  end
+end
