+++ app/views/api/v1/properties/index.json.jbuilder
@@ -1,0 +1,5 @@
+json.units do
+  json.array! @units do |unit|
+    json.partial! 'api/v1/shared/unit', unit: unit
+  end
+end
