+++ app/views/api/v1/shared/_unit.json.jbuilder
@@ -1,0 +1,9 @@
+json.id unit.code
+json.name unit.name
+json.type unit.type
+json.bedrooms unit.bedrooms
+json.bathrooms unit.bathrooms
+json.occupancy unit.occupancy
+json.position unit.position
+json.standardImages unit.standard_images
+json.preview_amount unit.preview_amount
