--- app/views/layouts/_header.html.erb
@@ -1,27 +1,0 @@
-<section class="site-header <%= content_for :header_class %>">
-    <nav class="nav-bar">
-        <div class="layout-container">
-            <div class="brand-container">
-                <a class="brand" href="/">
-                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 89 27">
-                        <g>
-                            <path d="M19.9 3.9c-1.4-0.1-2.7 0.2-4 0.8c0.6 0 1.2 0 1.8 0c1.4 0 2.4 0.9 3 2.2c0.5 1.2 0.4 2.3 0.2 3.6 c-0.4 2.2-1.1 4.3-2.1 6.2c-0.9 1.8-2 3.5-3.3 5C14.4 23 13 24.1 11.4 25c-1 0.6-2.1 1-3.2 1.4c-1.7 0.5-3.4 0.5-5-0.4 c-1-0.6-1.8-1.5-2.3-2.6c-0.7-1.7-1-3.4-0.9-5.2c0.1-1.7 0.4-3.3 1.2-4.9c1.1-2.2 2.3-4.2 4-6c0.9-1 1.8-2 2.9-2.8 c1.1-0.9 2.3-1.8 3.6-2.6c0.9-0.6 1.8-1 2.8-1.4c1.1-0.4 2.3-0.6 3.4-0.2c0.3 0.1 0.5 0.3 0.7 0.4c0.9 0.8 1.3 1.8 1.5 2.9 C20.1 3.7 20 3.8 19.9 3.9z M7.9 10.9C7.4 11.6 7 12 6.7 12.6c-1.2 1.9-2.3 3.8-3 6c-0.4 1.1-0.8 2.2-0.8 3.4C3 23.4 3.6 24 5 24.5 c0.8 0.2 1.5 0 2.2-0.3c1.9-0.8 3.5-2.1 4.9-3.6c2-2 3.6-4.3 4.9-6.9c0.7-1.3 1.1-2.7 1.3-4.2c0.1-0.5 0-1.1-0.2-1.6 c-0.2-0.8-0.8-1.2-1.6-1.2c-0.7 0-1.4 0.2-2 0.4c-1.2 0.5-2.3 1.2-3.3 2c-0.8 0.7-1.6 1.5-2.3 2.3c-0.7 0.8-1.2 1.7-1.9 2.6 C6.9 12.9 7.8 12.1 7.9 10.9z"/><path d="M57.5 13.6c0.1 0.3 0.1 0.6 0.3 0.8c0.6 1.1 0.3 2-0.4 2.8c-0.5 0.7-0.9 1.4-1.5 2c-0.3 0.4-0.4 0.8-0.5 1.2 c-0.2 0.7-0.1 1.3 0.5 1.7c0.5 0.3 1.2 0.3 1.8 0c0.5-0.2 0.9-0.6 1.3-0.8c0.2-0.1 0.5-0.2 0.8-0.3c0.2 0.8 0.2 1.4-0.5 2 c-1.1 1-2.1 2-3.3 2.8c-0.5 0.3-1 0.6-1.6 0.7c-1.3 0.4-2-0.2-2.1-1.5c0-1.1 0.1-2.1 0.4-3.2c0.1-0.3 0.2-0.6 0.1-1 c-0.2 0.2-0.4 0.3-0.5 0.5c-0.9 0.9-1.9 1.8-2.7 2.8c-0.6 0.7-1.3 1.2-2 1.6c-0.6 0.4-1.2 0.4-1.7-0.2c-1.2-1.3-1.9-2.7-2-4.5 c-0.1-1.8 0.6-3.2 1.8-4.4c1.6-1.5 3.3-2.9 5.3-4c1-0.5 2-1.1 3.1-1.5c0.8-0.3 1.6-0.4 2.4-0.6c0.8-0.2 1.3 0.2 1.7 0.9 c0.1 0.2 0.2 0.4 0.3 0.7c0.2 0.6 0.1 0.7-0.5 1c-0.1 0.1-0.2 0.1-0.3 0.2C57.6 13.4 57.6 13.5 57.5 13.6z M55.6 14.3 c-1.2 0.6-2.3 1.1-3.4 1.7c-1.4 0.7-2.8 1.5-4 2.6c-0.4 0.4-0.9 0.9-1.1 1.5c-0.3 0.6-0.6 1.2-0.3 2c0.3 0.7 0.7 0.9 1.3 0.5 c1.3-0.8 2.5-1.7 3.5-2.8c1.3-1.6 2.6-3.1 3.8-4.7C55.5 14.9 55.5 14.7 55.6 14.3z"/><path d="M74.9 20.5c0 0.6 0 1 0.1 1.4c0 0.4-0.2 0.6-0.4 0.9c-0.9 0.8-1.8 1.6-2.7 2.4c-0.6 0.5-1.3 0.8-2 1.1 c-0.5 0.2-0.9 0.2-1.4-0.1c-0.6-0.3-0.8-0.7-0.9-1.4c-0.2-1.8 0.6-3.2 1.4-4.7c0.5-0.9 1-1.8 1.5-2.7c-0.6 0.2-1.1 0.6-1.6 1 c-1.5 1.5-2.9 3.1-4.2 4.8c-0.4 0.6-0.9 1.2-1.4 1.8c-0.6 0.7-1.3 0.6-1.8-0.2c-0.5-0.9-0.8-1.9-0.8-3c0-0.7-0.1-1.4 0.1-2.1 c0.2-0.7 0.5-1.4 0.9-2.1c0.5-0.9 1.2-1.8 1.8-2.6c0.5-0.6 1.1-1.1 1.8-1.6c0.6-0.4 1.3-0.3 1.9-0.2c0.3 0.1 0.5 0.5 0.6 0.7 c-0.3 1.1-1.1 1.7-1.8 2.5c-0.7 0.8-1.3 1.7-2 2.5c-0.4 0.5-0.4 0.9-0.3 1.6c0.2-0.1 0.4-0.1 0.5-0.2c0.3-0.3 0.6-0.6 1-1 c1.5-1.6 3-3.2 4.5-4.7c1-0.9 2.1-1.7 3.2-2.4c1.2-0.8 1.8-0.5 2.2 0.9c0.3 1 0.2 1.9-0.6 2.8c-1.3 1.4-2.7 2.9-3.4 4.7 c-0.2 0.5-0.3 1 0.1 1.4c0.9 0.2 1.5-0.3 2.2-0.7C73.9 21.3 74.3 21 74.9 20.5z"/><path d="M86 20.7c0.3 0.4 0.7 0.4 1.1 0.2c0.5-0.3 0.9-0.6 1.4-0.9c0.1-0.1 0.2-0.1 0.4-0.2c0.3 1.3 0.2 1.4-0.7 2.1 c-1 0.7-2.1 1.2-3.3 1.2c-0.5 0-0.7 0.4-1 0.7c-0.6 0.6-1.2 1.1-1.9 1.6c-0.9 0.7-1.9 0.8-2.9 0.6c-0.6-0.1-1.2-0.5-1.6-0.9 c-1.6-1.9-2.3-4.6-1.1-7.1c1.4-2.9 3.7-4.9 6.7-6.2c1-0.4 1.9-0.5 2.9-0.1c0.5 0.2 0.7 0.6 0.9 1c-0.1 0.3-0.2 0.5-0.3 0.8 c1.2 0.8 1.3 1.4 1 2.9c-0.2 1.2-0.6 2.3-1.1 3.4C86.3 20.1 86.2 20.4 86 20.7z M84.4 15c0 0-0.1-0.1-0.1-0.1 c-0.8 0.5-1.6 0.9-2.3 1.5c-0.7 0.5-1.3 1.1-2 1.6c-1 1-1.3 2.3-1.5 3.6c-0.2 0.9 0.2 1.5 1.3 1.6c1.1 0.1 2-0.5 2.8-1.1 c0.3-0.2 0.3-0.6 0.2-0.9c-0.1-0.7-0.2-1.4-0.1-2c0.2-1.1 0.6-2.2 1.2-3.2C84.1 15.6 84.2 15.3 84.4 15z"/><path d="M35.6 20.6c-0.1 0.8-0.1 1.5 0.4 2.1c0.3 0.4 0.8 0.7 1.3 0.7c0.8 0 1.6-0.1 2.3-0.5c1.2-0.6 2.3-1.3 3.2-2.2 c0.6 1 0.6 1.3-0.2 2.1c-0.9 0.8-1.7 1.7-2.7 2.4c-1.2 0.9-2.6 1.5-4.2 1.5c-1.3 0-2.3-0.6-2.9-1.8c-0.8-1.5-0.9-3-0.7-4.7 c0.2-2.1 1.3-3.6 2.7-5c0.9-1 2-1.9 3.2-2.6c1.1-0.7 2.3-1.2 3.5-1.3c0.6-0.1 1.2 0.1 1.8 0.3c0.5 0.2 0.7 0.7 0.8 1.3 c0.1 1.2-0.4 2.2-1.2 3.1c-1 1.2-2.3 2.1-3.7 2.8C38 19.3 36.8 19.9 35.6 20.6z M40.7 13.6c-2.3 1.1-4.2 2.7-5 5.3 c0.2 0 0.2 0.1 0.3 0.1c1.7-0.7 3.1-1.8 4.3-3.2c0.3-0.4 0.5-0.9 0.8-1.4C41.2 14 41.1 13.8 40.7 13.6z"/><path d="M27.2 17.6c0-0.3 0-0.7-0.1-1.1c-0.3 0.1-0.5 0.1-0.7 0.3c-1 0.7-2 1.5-3 2.2c-0.5 0.4-1 0.9-1.1 1.6 c-0.3 1.5 0.9 2.9 2.4 3c1.2 0.1 2.3-0.4 3.3-1c1-0.6 1.9-1.2 2.9-1.8c0.2 0.4 0.1 0.8 0.1 1.2c0 0.3-0.2 0.6-0.4 0.8 c-1 0.9-2 1.9-3 2.7c-0.9 0.7-1.9 1.2-3.1 1.5c-1.1 0.2-2.1-0.1-2.9-1c-1-1-1.5-2.3-1.7-3.7c-0.1-1-0.2-1.9 0-2.9 c0.2-0.8 0.5-1.5 1-2c1.1-1.2 2.2-2.4 3.4-3.5c1.2-1.1 2.6-2.1 4.2-2.6c1.1-0.3 2.1 0.1 2.4 1c0.2 0.5 0.2 1.1-0.1 1.6 c-0.6 1.3-1.4 2.4-2.3 3.4C28.2 17.9 27.8 17.9 27.2 17.6z"/><path d="M47.2 2.4c-0.1 0.2-0.3 0.7-0.6 1c-1.3 1.5-2.8 2.9-4.5 4.1c-0.4 0.2-0.7 0.4-1.2 0.3c-0.5-0.1-0.7-0.2-0.7-0.7 c0-0.3 0.1-0.6 0.3-0.9c0.7-1.4 1.6-2.8 2.6-4C44 1.3 45 0.9 46.1 1.1C46.8 1.1 47.2 1.6 47.2 2.4z"/>
-                        </g>
-                    </svg>
-                </a>
-            </div>
-            <ul class="main-navigation">
-                <li><a href="/">Home</a></li>
-                <li><a href="/accommodations">Find a Rental</a></li>
-                <li><a href="/resources">Resources</a></li>
-                <li><a href="/contact">Contact</a></li>
-            </ul>
-        </div>
-    </nav>
-    <div class="hero <%= content_for :hero_class %>">
-        <div class="layout-container layout-centered padded-top padded-bottom">
-            <h1 class="<%= content_for :bottom_border %> heading-large"><%= yield(:hero_title) %></h1>
-            <% if content_for?(:lede) %><p class="lede"><%= yield :lede %></p><% end %>
-        </div>
-    </div>
-</section>
