--- app/views/layouts/_modernizr.html.erb
@@ -1,1 +1,0 @@
-<%= javascript_include_tag "modernizr" %>
