--- app/views/layouts/_scripts.html.erb
@@ -1,8 +1,0 @@
-<script src="//code.jquery.com/jquery-2.1.4.min.js"></script>
-<%= javascript_include_tag "moment" %>
-<%= javascript_include_tag "pikaday" %>
-<%= javascript_include_tag "pikaday.jquery" %>
-<%= javascript_include_tag "jquery.fancybox.pack.js" %>
-<%= javascript_include_tag "jquery.fancybox-media.js" %>
-
-<%= javascript_include_tag "application", params[:controller] %>
