--- app/views/layouts/application.html.erb
@@ -1,51 +1,0 @@
-<!doctype html>
-<!--[if IE 7]> <html class="no-js ie7 ie" lang="en"> <![endif]-->
-<!--[if IE 8]> <html class="no-js ie8 ie" lang="en"> <![endif]-->
-<!--[if gt IE 8]><!--> <html class="no-js" lang="en"> <!--<![endif]-->
-<head>
-    <meta charset="utf-8"/>
-    <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1" />
-
-    <title>Oc&eacute;ano Rentals</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta name="description" content="">
-    <meta name="author" content="">
-
-    <link rel="shortcut icon" type="image/ico" href="/favicon.ico" />
-    <link rel="apple-touch-icon" href="/apple-touch-icon.png">
-
-    <%= csrf_meta_tags %>
-
-    <%= render '/layouts/typekit_embed_code' %>
-    <%= render '/layouts/modernizr' %>
-
-    <%= stylesheet_link_tag("application") %>
-    <%= yield :stylesheets %>
-
-    <%= favicon_link_tag asset_path('favicon.png') %>
-</head>
-
-<body>
-    <!-- Social Icon SVG definitions -->
-    <svg xmlns="http://www.w3.org/2000/svg" style="display: none;">
-        <symbol id="twitter" viewBox="0 0 20 16">
-            <path fill="#ffffff" d="M.76 13.93c2.07.182 3.93-.3 5.6-1.59-1.754-.146-2.925-1.024-3.572-2.66.573.054 1.116.075 1.675-.078-1.89-.576-2.9-1.814-3.03-3.82.556.28 1.093.46 1.735.48-.84-.617-1.382-1.378-1.592-2.35-.21-.962-.083-1.88.418-2.753C4.078 3.587 6.684 4.942 9.9 5.162c-.012-.092-.017-.16-.03-.223-.36-1.998.92-3.966 2.892-4.393 1.342-.29 2.548.005 3.55.983.1.1.19.115.318.085.712-.165 1.393-.41 2.03-.768.06-.033.12-.06.207-.106-.293.907-.845 1.57-1.61 2.075.723-.066 1.407-.264 2.098-.51-.128.172-.238.362-.388.51-.442.444-.896.874-1.353 1.3-.087.08-.13.15-.13.27.04 1.7-.275 3.336-.996 4.877-.934 1.997-2.278 3.66-4.174 4.824-1.24.763-2.6 1.232-4.044 1.433-1.716.236-3.406.13-5.06-.406-.846-.275-1.643-.653-2.395-1.125-.014-.01-.023-.026-.055-.06" />
-        </symbol>
-        <symbol id="facebook" viewBox="0 0 12 21">
-            <path fill="#ffffff" d="M7.64 8.14h3.433c-.036.29-.067.563-.102.834-.108.834-.222 1.667-.325 2.5-.018.147-.07.188-.213.187-.868-.005-1.735-.003-2.602-.003h-.225v9H3.983V11.66H.96V8.144h3.02c0-.097-.002-.17 0-.245.004-.968-.034-1.94.023-2.905.093-1.558.75-2.807 2.175-3.565C6.88 1.057 7.645.91 8.43.92c.86.012 1.72.07 2.577.11.062.004.123.024.19.038v3.115h-.222c-.695.01-1.392.002-2.086.038-.752.04-1.202.452-1.24 1.2-.046.895-.01 1.793-.01 2.72" />
-        </symbol>
-        <symbol id="instagram" viewBox="0 0 113.4 113.4">
-            <path fill="#ffffff" d="M0 13.1v87.1c0 7.2 5.8 13.1 13.1 13.1h87.1c7.2 0 13.1-5.8 13.1-13.1V13.1c0-7.2-5.8-13.1-13.1-13.1H13.1C5.8 0 0 5.8 0 13.1zm78.4 4.3c0-2.4 1.9-4.4 4.4-4.4h13.1c2.4 0 4.4 2 4.4 4.4v13.1c0 2.4-2 4.4-4.4 4.4H82.7c-2.4 0-4.4-1.9-4.4-4.4V17.4zM56.6 34.8c12 0 21.8 9.7 21.8 21.8 0 12-9.7 21.8-21.8 21.8s-21.8-9.7-21.8-21.8c0-12 9.8-21.8 21.8-21.8zM13.1 47.9H23c-.7 2.8-1.1 5.7-1.1 8.7 0 19.2 15.6 34.8 34.8 34.8 19.2 0 34.8-15.6 34.8-34.8 0-3-.4-5.9-1.1-8.7h9.9v47.9c0 2.4-2 4.4-4.4 4.4H17.4c-2.4 0-4.4-1.9-4.4-4.4V47.9z"/>
-        </symbol>
-    </svg>
-
-    <%= render 'layouts/header' %>
-
-    <%= yield %>
-
-    <%= render 'layouts/footer' %>
-
-    <%= render 'layouts/scripts' %>
-
-</body>
-</html>
