--- app/views/pages/contact_thank_you.html.erb
@@ -1,11 +1,0 @@
-<% content_for :hero_title, "Thank you!".html_safe %> <!-- Hero/Page Title -->
-<% content_for :bottom_border, 'bottom-border' %> <!-- Does the header have a bottom border in the hero? -->
-
-<section class="padded-bottom main-content">
-    <div class="glass-background layout-container layout-centered padding-vertical">
-        <h1>Thank you for contacting us.</h1>
-        <p>We'll get back with you as soon as possible.</p>
-        <hr>
-        <a class="btn" href="/">Back to home</a>
-    </div>
-</section>
