--- app/views/pages/maps.html.erb
@@ -1,29 +1,0 @@
-<% content_for :hero_title, "Maps".html_safe %> <!-- Hero/Page Title -->
-<% content_for :hero_size, 'hero-small' %> <!-- What's the hero size for the header? -->
-<% content_for :bottom_border, 'bottom-border' %> <!-- Does the header have a bottom border in the hero? -->
-
-<section class="padded-bottom">
-    <div class="glass-background layout-container padding-vertical">
-        <div class="gw">
-            <nav class="subnav g md-one-third">
-                <ul>
-                    <li><a href="contact.html">Contact Us</a></li>
-                    <li><a class="current" href="maps.html">Maps</a></li>
-                    <li><a href="faq.html">FAQ</a></li>
-                </ul>
-            </nav>
-            <div class="subpage-content g md-two-thirds">
-                <h1 class="heading-large">Maps</h1>
-                <p>For help getting to and around Rocky Point, here are some maps.</p>
-                <h2 class="heading-medium top-spacing">Overview</h2>
-                <a href="/files/maps/oceano_map_overview.jpg" alt="Standalone overview map">
-                    <%= image_tag "/files/maps/oceano_map_overview.jpg", :class => "map-image" %>
-                </a>
-                <h2 class="heading-medium top-spacing">Details</h2>
-                <a href="/files/maps/oceano_map_detail.jpg" alt="Standalone detail map">
-                    <%= image_tag "/files/maps/oceano_map_detail.jpg", :class => "map-image" %>
-                </a>
-            </div>
-        </div>
-    </div>
-</section>
