--- app/views/pages/resources.html.erb
@@ -1,71 +1,0 @@
-<% content_for :hero_title, "Resources".html_safe %> <!-- Hero/Page Title -->
-<% content_for :bottom_border, 'bottom-border' %> <!-- Does the header have a bottom border in the hero? -->
-
-<section class="padded-top-50 padded-bottom">
-    <div class="glass-background layout-container margin-vertical padding-vertical">
-        <div class="gw large-gutter">
-            <div class="g sm-one-third">
-                <h2 class="heading-medium-small up">The Essentials</h2>
-                <p>Things you need before you go to Rocky Point</p>
-            </div>
-            <div class="g sm-one-third">
-                <h2 class="heading-medium-small up">Home Insurance</h2>
-                <p>Things you need before you go to Rocky Point</p>
-                <p><a href="https://sb.iigins.com/mexico-home-owners/quick-quote/quote_step_1.mhtml?aff_id=3732">Get Home Insurance &gt;</a></p>
-            </div>
-            <div class="g sm-one-third">
-                <h2 class="heading-medium-small up">Car Insurance</h2>
-                <p>Things you need before you go to Rocky Point</p>
-                <p><a href="https://sb.iigins.com/quote/?aff_id=3732&agtdst=DEKACF&office_code=;">Get Car Insurance &gt;</a></p>
-            </div>
-            <div class="g sm-one-third">
-                <h2 class="heading-medium-small up">Tidal Forecast</h2>
-                <p>When to expect high and low tide for the area.</p>
-                <p><a href="http://www.tide-forecast.com/locations/Puerto-Penasco-Sonora-Mexico/tides/latest">View Forecast &gt;</a></p>
-            </div>
-            <div class="g sm-one-third">
-                <h2 class="heading-medium-small up">Property List</h2>
-                <p><a href="/properties">Compare Properties &gt;</a></p>
-            </div>
-        </div>
-    </div>
-</section>
-
-<section class="padding-vertical">
-    <div class="glass-background layout-container layout-centered margin-vertical padding-vertical">
-        <h1 class="heading-medium-small up alt">Enjoy Your Stay</h1>
-        <h2 class="heading-subheader">Nice extras to help you stay connected and enjoy your trip.</h2>
-    </div>
-</section>
-
-<section class="padded-bottom">
-    <div class="glass-background layout-container padding-vertical">
-        <div class="gw large-gutter">
-            <div class="g sm-one-half">
-                <h2 class="heading-medium">Massage</h2>
-                <p><strong class="slate">Full Body</strong> - 1.5 hrs, $75<br />
-                <strong class="slate">Full Body</strong> - 1 hr, $55<br />
-                <strong class="slate">Upper Body</strong> - 35 min, $35</p>
-
-                <p>Senior citizens 10% off</p>
-
-                <ol>
-                    <li>Swedish</li>
-                    <li>Shiatsu</li>
-                    <li>Deep shiatsu</li>
-                    <li>Reflexology</li>
-                    <li>Elastisiri massage</li>
-                    <li>Dren linfatic System</li>
-                </ol>
-
-                <p>If you were getting a Full Body massage for 1hour and 30 minutes you will receive all All 6 types of massages.</p>
-
-                <p>If you were getting the Full Body Massage for 1 hour you would receive massage types 1, 2, 3</p>
-
-                <p>If you were getting the Upper Body for 35 minutes you would receive massage types 1, 2</p>
-
-                <p>A licensed massage therapist who only uses 100% Aztec oils with healing ingredients applies these massages.</p>
-            </div>
-        </div>
-    </div>
-</section>
