--- app/views/pages/testimonials.html.erb
@@ -1,48 +1,0 @@
-<% content_for :hero_title, "Testimonials".html_safe %> <!-- Hero/Page Title -->
-<% content_for :hero_size, 'hero-small' %> <!-- What's the hero size for the header? -->
-<% content_for :bottom_border, 'bottom-border' %> <!-- Does the header have a bottom border in the hero? -->
-
-<section class="padded-bottom">
-    <div class="glass-background layout-container padding-vertical margin-vertical">
-        <ul class="testimonials normal gw large-gutter">
-            <li class="g sm-one-half">
-                <blockquote>
-                    <p class="persons-words">&ldquo;We loved the house &amp; location! We have always stayed at the condos on Sandy Beach, and loved this location so much more. So quiet and peaceful and quick access to the fish market. Will likely rent this unit in the future.&rdquo;</p>
-                    <span class="persons-name">Tucson, AZ</span>
-                </blockquote>
-            </li>
-            <li class="g sm-one-half">
-                <blockquote>
-                    <p class="persons-words">&ldquo;We had an awesome girls trip. The house was right on the beach just like we wanted it to be. It served all our needs and it was very close to downtown like we were hoping for. Overall we had a great time and we loved our stay at the Casita de Sueños.&rdquo;</p>
-                    <span class="persons-name">San Bernardino, CA</span>
-                </blockquote>
-            </li>
-        </ul>
-    </div>
-    <div class="testimonials featured">
-        <div class="glass-background layout-container">
-            <blockquote>
-                <p class="persons-words">&ldquo;Our stay was for 5 days. The house was well maintained and there were no unexpected problems. There was no need to call for any maintenance.
-Great little place for 4. Very cozy. Everyone was very satisfied with the simple design of the house and the great beds in the bedroom. Such soft and plush beds. They really made the vacation even better.
-The people at Oceano were great but what was nice was not having to fiddle with any problems in the home you could see they took great care to make the home a rock solid vacation home. Close to everything in town.&rdquo;</p>
-                <span class="persons-name">Sierra Vista, AZ</span>
-            </blockquote>
-        </div>
-    </div>
-    <div class="glass-background layout-container padding-vertical margin-vertical">
-        <ul class="testimonials normal gw large-gutter">
-            <li class="g sm-one-half">
-                <blockquote>
-                    <p class="persons-words">&ldquo;Quiet. Relaxing. Very well done and decorated condo in beautiful surroundings. The staff was superb.&rdquo;</p>
-                    <span class="persons-name">Silverthorne, CO</span>
-                </blockquote>
-            </li>
-            <li class="g sm-one-half">
-                <blockquote>
-                    <p class="persons-words">&ldquo;Great time. Clean place with awesome beach front view. Would stay there again.&rdquo;</p>
-                    <span class="persons-name">Phoenix, AZ</span>
-                </blockquote>
-            </li>
-        </ul>
-    </div>
-</section>
