--- app/views/pages/trip_cancellation_insurance.html.erb
@@ -1,13 +1,0 @@
-<% content_for :hero_title, "Trip Cancellation Insurance".html_safe %> <!-- Hero/Page Title -->
-<% content_for :hero_size, 'hero-small' %> <!-- What's the hero size for the header? -->
-<% content_for :bottom_border, 'bottom-border' %> <!-- Does the header have a bottom border in the hero? -->
-
-<section class="padded-top padded-bottom">
-    <div class="layout-container layout-centered">
-        <h1 class="heading-medium">Optional Financial Protection</h1>
-    </div>
-    <div class="layout-container-smaller">
-      <p>Your online booking includes the option of Trip Cancellation Insurance. Trip Cancellation Insurance provides financial protection for your vacation rental. You have the option to decline this insurance by checking the "decline box" on the Summary of Charges page during the on-line booking process. The policy details are available in a PDF format to view and or print. Your credit card will be charged by CSA Travel Protection. Each policy has a 10 Day Free Look Period, beginning from the date you receive your policy, in which you can decide to cancel this option. CSA will e-mail a confirmation letter and certificate of insurance to the e-mail address you provided.</p>
-      <p>Dowload this document for more information on why you should purchase Trip Cancellation Insurance: <a href="/files/pdfs/why_buy_trip_insurance.pdf" target="_blank">Why_Buy_Trip_Insurance.pdf</a> <em>(108K - Adobe Acrobat required)</em></p>
-    </div>
-</section>
