--- app/views/properties/_rate_panel.html.erb
@@ -1,134 +1,0 @@
-<aside class="request-booking md-one-third">
-<%- if @lookup && @available %>
-    <header class="booking-header">
-        <h2>$<%= '%.0f' % @nightly_rate %> <span class="term">Per Night</span></h2>
-    </header>
-    <%= form_tag("/accommodations/#{@id}", {method: 'get', class: 'gw no-gutter booking-details'}) do %>
-    <table>
-        <tr class="date-select">
-            <td class="g sm-three-eighths">
-                <label for="start_date">Arrive</label>
-                <%= text_field_tag "start_date", @start_date, class: 'no-borders datepicker', placeholder: 'mm/dd/yyyy' %>
-            </td>
-            <td class="g sm-three-eighths">
-                <label for="end_date">Depart</label>
-                <%= text_field_tag "end_date", @end_date, class: 'no-borders datepicker', placeholder: 'mm/dd/yyyy' %>
-            </td>
-            <td class="g sm-two-eighths">
-                <label for="guests">Guests</label>
-                <div class="select-style">
-                <%= select_tag "guests", options_for_select(@guest_amount_list, @guests) %>
-                </div>
-            </td>
-        </tr>
-        <tr class="full-width-container">
-            <td>
-                <input class="btn no-min-width" type="submit" value="Search Again">
-            </td>
-        </tr>
-        <tr class="details first">
-            <td class="g two-thirds">$<%=  '%.0f' % @nightly_rate %> x <%= @length_of_stay %> nights</td>
-            <td class="g one-third">$<%= @base_amount %></td>
-        </tr>
-        <tr class="details">
-            <td class="g two-thirds">Tax</td>
-            <td class="g one-third">$<%= @tax_amount %></td>
-        </tr>
-        <% @fees.each do |fee| %>
-        <tr class="details">
-            <td class="g two-thirds"><%= fee.name %></td>
-            <td class="g one-third">$<%= '%.2f' % fee.amount %></td>
-        </tr>
-        <% end %>
-        <tr class="details">
-            <td class="g two-thirds highlight">Total</td>
-            <td class="g one-third highlight">$<%= '%.2f' % @total_amount %></td>
-        </tr>
-        <tr class="details">
-            <td class="g two-thirds"><a href="/trip-cancellation-insurance" target="_blank">Optional Trip Cancellation Insurance</a></td>
-            <td class="g one-third">$<%= '%.2f' % ((@base_amount.to_f * 1.12) * 0.065) %></td>
-        </tr>
-        <tr class="details">
-            <td class="g two-thirds highlight">Total w/ Insurance</td>
-            <td class="g one-third highlight">$<%= '%.2f' % (((@base_amount.to_f * 1.12) * 0.065) + @total_amount.to_f) %></td>
-        </tr>
-        <tr class="full-width-container">
-            <td>
-                <a href="https://bookings-oceano-rentals.escapia.com/Booking/RateDetails/<%= @booking_id %>?arrive=<%= @start_date %>&depart=<%= @end_date %>&adults=<%= @guests %>" class="btn request-booking-link">
-                    Request Booking
-                </a>
-            </td>
-        </tr>
-        <tr class="full-width-container">
-            <td>
-                <span class="note">The total above is an estimate and includes rent, taxes, and additional fees.</span>
-            </td>
-        </tr>
-    </table>
-    <% end %>
-<%- elsif @lookup && !@available %>
-    <header class="booking-header">
-        <h2>Availability and Rate</h2>
-    </header>
-    <%= form_tag("/accommodations/#{@id}", {method: 'get', class: 'gw no-gutter booking-details'}) do %>
-    <table>
-        <tr class="date-select">
-        <td class="g sm-three-eighths">
-            <label for="start_date">Arrive</label>
-            <%= text_field_tag "start_date", @start_date, class: 'no-borders datepicker', placeholder: 'mm/dd/yyyy' %>
-        </td>
-        <td class="g sm-three-eighths">
-            <label for="end_date">Depart</label>
-            <%= text_field_tag "end_date", @end_date, class: 'no-borders datepicker', placeholder: 'mm/dd/yyyy' %>
-        </td>
-        <td class="g sm-two-eighths">
-            <label for="guests">Guests</label>
-            <div class="select-style">
-            <%= select_tag "guests", options_for_select(@guest_amount_list, @guests) %>
-            </div>
-        </td>
-        </tr>
-        <tr class="full-width-container">
-            <td class="g">This property is unavailable for the chosen dates.</td>
-        </tr>
-        <tr class="full-width-container">
-        <td>
-            <input class="btn no-min-width" type="submit" value="Check Availability">
-        </td>
-        </tr>
-    </table>
-    <% end %>
-<%- else %>
-    <header class="booking-header">
-        <h2>Availability and Rate</h2>
-    </header>
-    <%= form_tag("/accommodations/#{@id}", { method: 'get', class: 'gw no-gutter booking-details' }) do %>
-    <table>
-        <tr class="date-select">
-        <td class="g sm-three-eighths">
-            <label for="start_date">Arrive</label>
-            <%= text_field_tag "start_date", @start_date, class: 'no-borders datepicker', placeholder: 'mm/dd/yyyy' %>
-        </td>
-        <td class="g sm-three-eighths">
-            <label for="end_date">Depart</label>
-            <%= text_field_tag "end_date", @end_date, class: 'no-borders datepicker', placeholder: 'mm/dd/yyyy' %>
-        </td>
-        <td class="g sm-two-eighths">
-            <label for="guests">Guests</label>
-            <div class="select-style">
-            <%= select_tag "guests", options_for_select(@guest_amount_list, @guests) %>
-            </div>
-        </td>
-        </tr>
-        <tr class="full-width-container">
-            <td class="g">Check availablilty to view rates and fees.</td>
-        </tr>
-        <tr class="full-width-container">
-        <td>
-            <input class="btn no-min-width" type="submit" value="Check Availability">
-        </td>
-        </tr>
-    </table>
-    <% end %>
-<% end %>
-</aside>
