--- app/views/properties/index.html.erb
@@ -1,60 +1,0 @@
-<% content_for :hero_title, "Find a Rental".html_safe %> <!-- Hero/Page Title -->
-<% content_for :lede, "Search for a rental at Rocky Point".html_safe %>
-<section class="padded-top padded-bottom">
-  <%= render 'shared/search_form' %>
-  <%- if @units.empty? %>
-    <div class="layout-container-smaller padded-bottom layout-centered">
-        <h1>No available rental properties found</h1>
-        <p>Please adjust your search and try again.</p>
-    </div>
-    <%- else %>
-    <div class="glass-background layout-container">
-        <ul class="properties-list gw small-gutter">
-            <%- @units.each do |unit| %>
-            <li class="property g sm-one-half md-one-whole">
-
-                <%
-                full_link = accommodation_url(unit.code) + "?start_date=#{@start_date}&end_date=#{@end_date}&guests=#{@guests}"
-                %>
-
-                <a href="<%= full_link %>"><img class="property-thumb" src="<%= unit.standard_images[0][:url] %>"></a>
-                <div class="details">
-                    <div class="gw small-gutter">
-                        <div class="g md-three-eighths">
-                            <h2><a href="<%= full_link %>"><%= unit.name %></a></h2>
-                            <p class="location"><%= unit.address.street %> <br><%= unit.address.city %>, <%= unit.address.country %></p>
-                        </div>
-                        <div class="g md-three-eighths">
-                            <ul class="offerings">
-                                <li>
-                                    <i class="icon icon-bath"></i>
-                                    <span class="detail"><%= unit.bathrooms %> bath</span>
-                                </li>
-                                <li>
-                                    <i class="icon icon-bed"></i>
-                                    <span class="detail"><%= unit.bedrooms %> beds</span>
-                                </li>
-                                <li>
-                                    <i class="icon icon-guests"></i>
-                                    <span class="detail"><%= unit.occupancy %> guests</span>
-                                </li>
-                            </ul>
-                        </div>
-                        <div class="g md-two-eighths">
-                            <span class="price"></span>
-                            <span class="type"><%= unit.type %></span>
-                        </div>
-                    </div>
-                </div>
-            </li>
-            <%- end %>
-
-        </ul>
-
-        <div class="pagination pagination-centered">
-        <%= will_paginate @units, previous_label: '<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16.3 32.6"><polygon points="16.3,32.6 0,16.3 16.3,0"/></svg>',
-                                  next_label: '<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16.3 32.6"><polygon points="0,0 16.3,16.3 0,32.6"/></svg>' %>
-        <%- end %>
-        </div>
-    </div>
-</section>
