--- app/views/properties/properties.html.erb
@@ -1,54 +1,0 @@
-<% content_for :hero_title, "Find a Rental".html_safe %> <!-- Hero/Page Title -->
-<% content_for :lede, "Search for a rental at Rocky Point".html_safe %>
-<section id="property-list" class="padded-top padded-bottom">
-  <div class="glass-background padded-top-40 margin-vertical">
-    <%= render 'shared/search_form' %>
-    <%- if @units.empty? %>
-      <div class="layout-container-smaller padded-bottom layout-centered">
-        <h1>No available rental properties found</h1>
-        <p>Please adjust your search and try again.</p>
-      </div>
-    <%- else %>
-      <div class="layout-container">
-
-        <table id="properties-table">
-          <colgroup>
-            <col style="width: 20%"/>
-            <col style="width: 19%"/>
-            <col style="width: 15%"/>
-            <col style="width: 8%"/>
-            <col style="width: 8%"/>
-            <col style="width: 11%"/>
-            <col style="width: 8%"/>
-            <col style="width: 11%"/>
-          </colgroup>
-          <thead>
-            <th>Preview</th>
-            <th>Property</th>
-            <th>Location</th>
-            <th>Beds</th>
-            <th>Baths</th>
-            <th>Occupancy</th>
-            <th>Beach</th>
-            <th>Type</th>
-          </thead>
-          <tbody>
-            <% @units.each do |unit| %>
-              <% full_link = accommodation_url(unit.code) + "?start_date=#{@start_date}&end_date=#{@end_date}&guests=#{@guests}" %>
-              <tr class="property">
-                <td><a href="<%= full_link %>"><img src="<%= fetch_image unit %>"/></a></td>
-                <td><a href="<%= full_link %>" class="property__name"><%= fetch_name unit %></a></td>
-                <td><%= unit.address.street %></td>
-                <td><%= unit.bedrooms %></td>
-                <td><%= unit.bathrooms %></td>
-                <td class="property__occupancy"><%= unit.occupancy %></td>
-                <td><%= t unit.beachfront.to_s %></td>
-                <td><%= unit.type %></td>
-              </tr>
-            <% end %>
-          </tbody>
-        </table>
-      <%- end %>
-    </div>
-  </div>
-</section>
