--- app/views/properties/show.html.erb
@@ -1,176 +1,0 @@
-<script type="text/javascript" src="//maps.googleapis.com/maps/api/js?sensor=false"></script>
-
-<% content_for :hero_title, @property_title.html_safe %>
-<% content_for :lede, @property_subtitle.html_safe %>
-
-<section>
-  <div class="glass-background layout-container">
-    <div class="md-two-thirds">
-      <ul class="offerings">
-        <li>
-          <i class="icon icon-bath slate"></i>
-          <span class="detail slate"><%= @unit.bathrooms %> Bath</span>
-        </li>
-        <li>
-          <i class="icon icon-bed slate"></i>
-          <span class="detail slate"><%= @unit.bedrooms %> Beds</span>
-        </li>
-        <li>
-          <i class="icon icon-guests slate"></i>
-          <span class="detail slate"><%= @unit.occupancy %> Guests</span>
-        </li>
-        <li>
-          <i class="icon icon-apartment slate"></i>
-          <span class="detail slate"><%= @unit.type.capitalize %></span>
-        </li>
-      </ul>
-    </div>
-
-    <%#======================================================================%>
-
-    <%= render 'rate_panel' %>
-
-    <div class="property-details md-two-thirds">
-
-      <div class="image-grouping" id="image-lightbox">
-        <div class="row">
-        <%- if @large_images.count > 0 %>
-          <article class="one-whole md-two-thirds tile">
-            <a rel="image-gallery" href="<%= @large_images[0][:url] %>" title="<%= @large_images[0][:title] %>" class="fancybox contain" style="background-image: url('<%= @standard_images[0][:url] %>');"></a>
-          </article>
-        <%- end %>
-        <%- if @large_images.count > 1 %>
-          <article class="one-whole md-one-third tile">
-            <a rel="image-gallery" href="<%= @large_images[1][:url] %>" title="<%= @large_images[1][:title] %>" class="fancybox contain" style="background-image: url('<%= @standard_images[1][:url] %>');"></a>
-          </article>
-        <%- end %>
-        </div>
-        <div class="row">
-        <%- if @large_images.count > 2 %>
-          <article class="one-whole md-one-third tile">
-            <a rel="image-gallery" href="<%= @large_images[2][:url] %>" title="<%= @large_images[2][:title] %>" class="fancybox contain" style="background-image: url('<%= @standard_images[2][:url] %>');"></a>
-          </article>
-        <%- end %>
-        <%- if @large_images.count > 3 %>
-          <article class="one-whole md-one-third tile">
-            <a rel="image-gallery" href="<%= @large_images[3][:url] %>" title="<%= @large_images[3][:title] %>" class="fancybox contain" style="background-image: url('<%= @standard_images[3][:url] %>');"></a>
-          </article>
-        <%- end %>
-
-          <% for i in 5..@standard_images.count do %>
-            <a rel="image-gallery" href="<%= @large_images[i-1][:url] %>" title="<%= @large_images[i-1][:title] %>" class="fancybox hidden"></a>
-          <% end %>
-
-          <article class="one-whole md-one-third tile">
-            <a href="#" class="contain black placehold-all-images open-lightbox">
-              <div class="over">
-                <h2>See All <%= @standard_images.count %> Photos</h2>
-              </div>
-            </a>
-          </article>
-        </div>
-      </div>
-
-      <div class="gw">
-        <div class="g">
-          <%
-            unless @videos.empty? || @videos[0][:formats][:url].empty?
-              url = @videos[0][:formats][:url]
-          %>
-
-          <p><a class="fancybox-media btn video-tour-btn" title="Video Tour" href="<%= url %>" target="_blank">View a video tour</a></p>
-          <% end %>
-        </div>
-      </div>
-
-
-      <div class="gw">
-        <div class="g md-one-fourth">
-          <h2 class="heading-two alt">Description</h2>
-        </div>
-        <div class="g md-three-fourths">
-          <p><%= @unit.descriptions[:text][0][:description] %></p>
-        </div>
-      </div>
-
-      <hr>
-
-      <div class="gw">
-        <div class="g md-one-fourth">
-          <h2 class="heading-two alt">Property Details</h2>
-        </div>
-        <div class="g md-three-fourths">
-          <ul class="gw no-gutter no-padding">
-            <li class="g md-one-half padded-right"><p><b class="slate">Property Type:</b> <%= @unit.type.capitalize %></p></li>
-            <li class="g md-one-half padded-right"><p><b class="slate">Bedrooms:</b> <%= @unit.bedrooms %></p></li>
-            <li class="g md-one-half padded-right"><p><b class="slate">Bathrooms:</b> <%= @unit.bathrooms %></p></li>
-            <li class="g md-one-half padded-right"><p><b class="slate">Floors:</b> <%= @unit.num_floors %></p></li>
-            <li class="g md-one-half padded-right"><p><b class="slate">Occupancy:</b> <%= @unit.occupancy %></p></li>
-          </ul>
-        </div>
-      </div>
-
-      <hr>
-      <div class="gw">
-        <div class="g md-one-fourth">
-          <h2 class="heading-two alt">Amenities</h2>
-        </div>
-        <div class="g md-three-fourths">
-          <ul class="gw no-gutter no-padding">
-            <% @amenities.each do |amenity| %>
-            <li class="g md-one-half padded-right"><p><%= amenity.to_s.titleize %></p></li>
-            <% end %>
-          </ul>
-        </div>
-      </div>
-    </div>
-  </div>
-</section>
-
-<%- @unit.reviews.sample(3).compact.tap do |reviews| %>
-<%- break if reviews.empty? %>
-<section class="padded-top padded-bottom">
-  <div class="glass-background layout-container padding-vertical">
-    <div class="md-two-thirds">
-      <h2 class="heading-medium bottom-spacing-large">Testimonies</h2>
-      <ul class="testimonials testimonials-paginated normal gw large-gutter">
-        <%- reviews.each do |review| %>
-        <li class="g sm-one-whole">
-          <blockquote>
-            <p class="persons-words">“<%= review[:comment].html_safe %></p>
-            <%- unless review[:city].empty? || review[:state].empty? %>
-            <span class="persons-name"><%= review[:city] %>, <%= review[:state] %></span>
-            <% end %>
-          </blockquote>
-        </li>
-        <%- end %>
-      </ul>
-    </div>
-  </div>
-</section>
-<%- end %>
-
-<div id="map-container">
-  <%
-    @map_latitude = @unit.position[:latitude]
-    @map_longitude = @unit.position[:longitude]
-  %>
-  <div id="map-canvas" data-lat="<%= @map_latitude %>" data-lon="<%= @map_longitude %>"></div>
-</div>
-
-<section class="padded-top padded-bottom related-places">
-  <div class="glass-background layout-container padding-vertical">
-    <h2 class="heading-medium bottom-spacing-large">Other Places</h2>
-    <ul class="gw large-gutter no-padding">
-    <%- @random_units.each do |unit| %>
-      <li class="g md-one-third">
-        <a href="<%= accommodation_url(unit.code) %>">
-          <img src="<%= unit.standard_images[0][:url] %>">
-          <h2 class="top-spacing slate"><%= unit.name %></h2>
-          <p><%= ellipsis(unit.long_description, length: 50) %></p>
-        </a>
-      </li>
-    <%- end %>
-    </ul>
-  </div>
-</section>
