--- app/views/shared/_search_form.html.erb
@@ -1,42 +1,0 @@
-<div class="layout-container">
-  <div class="glass-background property-search">
-    <%= form_tag("/accommodations", {method: 'get', class: 'gw no-gutter'}) do %>
-      <fieldset class="g md-one-fifth">
-        <label class="is-visually-hidden">Where are you going?</label>
-        <div class="select-style">
-          <%- areas_options = [['Where are you going?', '-', {disabled: 'disabled'}], ['Any', 'all']] + areas.map { |v| [v[:name], v[:key]] } %>
-          <%= select_tag "area", options_for_select(areas_options, @area) %>
-        </div>
-      </fieldset>
-      <fieldset class="g sm-one-third md-one-eighth">
-        <%= text_field_tag "start_date", @start_date, class: 'no-borders datepicker', placeholder: 'Arrive' %>
-      </fieldset>
-      <fieldset class="g sm-one-third md-one-eighth">
-        <%= text_field_tag "end_date", @end_date, class: 'no-borders datepicker', placeholder: 'Depart' %>
-      </fieldset>
-      <fieldset class="g sm-one-third md-one-tenth">
-        <label class="is-visually-hidden" for="guests">Guests</label>
-        <div class="select-style">
-          <%- guest_options = [['Guests', '-', {disabled: 'disabled'}], ['Any', 'all']] + (1..10).map { |v| [v, v] } %>
-          <%= select_tag "guests", options_for_select(guest_options, @guests) %>
-        </div>
-      </fieldset>
-      <fieldset class="g sm-one-third md-one-seventh">
-        <div class="select-style">
-          <%- beachfront_options = [['Beachfront', '-', { disabled: 'disabled' }], ['Any', 'all'], ['Yes', true], ['No', false]] %>
-          <%= select_tag 'beachfront', options_for_select(beachfront_options, @beachfront) %>
-        </div>
-      </fieldset>
-      <fieldset class="g sm-one-whole md-one-eighth">
-        <label class="is-visually-hidden" for="sort">Sort by</label>
-        <div class="select-style">
-          <%- sort_options = [['Sort by', '-', {disabled: 'disabled'}]] + sort_by %>
-          <%= select_tag "sort", options_for_select(sort_options, @sort) %>
-        </div>
-      </fieldset>
-      <footer class="g md-two-elevenths">
-        <input class="btn btn--green no-min-width max-width" type="submit" value="Search">
-      </footer>
-    <%- end %>
-  </div>
-</div>
